fix: split course crud into smaller component
